# Optimize Wallet Load (3.6x faster)