--- a/WalletWasabi/Blockchain/Analysis/Clustering/Cluster.cs
+++ b/WalletWasabi/Blockchain/Analysis/Clustering/Cluster.cs
@@ -9,7 +9,9 @@ namespace WalletWasabi.Blockchain.Analysis.Clustering
 {
 	public class Cluster : NotifyPropertyChangedBase, IEquatable<Cluster>
 	{
+		private object Lock { get; }
 		private List<SmartCoin> Coins { get; set; }
+		private HashSet<SmartCoin> CoinsSet { get; set; }
 
 		private SmartLabel _labels;
 
@@ -28,7 +30,9 @@ namespace WalletWasabi.Blockchain.Analysis.Clustering
 
 		public Cluster(IEnumerable<SmartCoin> coins)
 		{
+			Lock = new object();
 			Coins = coins.ToList();
+			CoinsSet = Coins.ToHashSet();
 			Labels = SmartLabel.Merge(Coins.Select(x => x.Label));
 		}
 
@@ -36,18 +40,29 @@ namespace WalletWasabi.Blockchain.Analysis.Clustering
 
 		public void Merge(IEnumerable<SmartCoin> coins)
 		{
-			var insertPosition = 0;
-			foreach (var coin in coins.ToList())
+			lock (Lock)
 			{
-				if (!Coins.Contains(coin))
+				var insertPosition = 0;
+				foreach (var coin in coins.ToList())
 				{
-					Coins.Insert(insertPosition++, coin);
+					if (CoinsSet.Add(coin))
+					{
+						Coins.Insert(insertPosition++, coin);
+					}
+					coin.Clusters = this;
 				}
-				coin.Clusters = this;
+				if (insertPosition > 0) // at least one element was inserted
+				{
+					Labels = SmartLabel.Merge(Coins.Select(x => x.Label));
+				}
 			}
-			if (insertPosition > 0) // at least one element was inserted
+		}
+
+		public IEnumerable<SmartCoin> GetCoins()
+		{
+			lock (Lock)
 			{
-				Labels = SmartLabel.Merge(Coins.Select(x => x.Label));
+				return Coins.ToList();
 			}
 		}
 
@@ -59,15 +74,18 @@ namespace WalletWasabi.Blockchain.Analysis.Clustering
 
 		public override int GetHashCode()
 		{
-			int hash = 0;
-			if (Coins != null)
+			lock (Lock)
 			{
-				foreach (var coin in Coins)
+				int hash = 0;
+				if (Coins is { })
 				{
-					hash ^= coin.GetHashCode();
+					foreach (var coin in Coins)
+					{
+						hash ^= coin.GetHashCode();
+					}
 				}
+				return hash;
 			}
-			return hash;
 		}
 
 		public static bool operator ==(Cluster x, Cluster y)
@@ -91,7 +109,10 @@ namespace WalletWasabi.Blockchain.Analysis.Clustering
 				}
 				else
 				{
-					return x.Coins.SequenceEqual(y.Coins);
+					lock (x.Lock)
+					{
+						return x.Coins.SequenceEqual(y.Coins);
+					}
 				}
 			}
 		}