As these two PRs modify the exact same code, this PR merges them together.

Closes https://github.com/zkSNACKs/WalletWasabi/pull/3080
Closes https://github.com/zkSNACKs/WalletWasabi/pull/3129

Measuring my own wallet load 3-3 times in release:

```
master: 21.59s, 21.71s, 22.17s
this pr: 18.06s, 17.97s, 18.26s
```