@nopara73 2020-02-23
With `SetEqual` (https://github.com/zkSNACKs/WalletWasabi/pull/3132/commits/de28812157e954052039fcc6aeb61ef557b8db73): `14.93, 14.51, 15.11`