@nopara73 2020-02-23
With `HashSet` instead of `FirstOrDefault` change (https://github.com/zkSNACKs/WalletWasabi/pull/3132/commits/a648f3c60060a17e8834efa102c9613fcc142f36): 7.44, 6.26, 6.97