@nopara73 2020-02-23
With `OutPoint` magic change (https://github.com/zkSNACKs/WalletWasabi/pull/3132/commits/7398cb0e7965495f5903039a485ec965846533a7): `5.85, 6.11, 5.66`