@yahiheb 2020-02-24
```
2020-02-24 14:56:42 WARNING     Wallet (101)    System.NullReferenceException: Object reference not set to an instance of an object.
   at WalletWasabi.Gui.Wallet.DesperateDequeueAllCoinsAsync() in C:\Users\RESEAU\Desktop\WalletWasabi\WalletWasabi.Gui\Wallet.cs:line 111
   at WalletWasabi.Gui.Wallet.TryDesperateDequeueAllCoinsAsync() in C:\Users\RESEAU\Desktop\WalletWasabi\WalletWasabi.Gui\Wallet.cs:line 93
```

I get this warning on this branch but I don't get it on master.