@molnard 2020-02-24
@yahiheb you are on the multi-wallet PR branch. Here we don't have Wallet.cs