@yahiheb 2020-02-24
> @yahiheb you are on the multi-wallet PR branch. Here we don't have Wallet.cs

My bad. I don't know how but I was on a branch named `clusterlockhashset` with different commits and I had `Wallet.cs` file.